Fixed issue #1: Small screen browsing not working
Fixed the stylesheet to specify default values that will be used on
small screens first, then on medium screens, then on larger screens.
This order is advised by W3 Schools.

It’s been tested by changing the browser window size on a laptop.
Hopefully, it will work on a mobile phone browser too.