Fixed error in submit button and its execution
Submit button was working by magic.  The HTML ended with a question
mark (typo!) instead of a closing pointy bracket.  When the typo was
corrected, the program stopped working because the click event wasn’t
being handled properly, with the e and the e.preventDefault and the
return false.  Now it is.  Thanks, Colin!