Added animated loading icon plus GitHub link plus more images
Added animated loading icon.  The icon is the Buffer logo rendered as
an SVG image; the image and its animation come courtesy of
https://cssanimation.rocks/buffer/.
The positioning of the animated loading icon comes courtesy of
http://stackoverflow.com/questions/8639383/how-do-i-center-an-svg-in-a-d
iv, with a bit of extra jiggery-pokery by me.
Added link to GitHub repository.
Added more images of the lovely Buffer people and the lovely WAGs and
HABs.